Introduce dungeon.py and treasure.py to simplify world/action layer
Extract dungeon predicates (defeated_monsters, defeated_dungeon,
blocking_dragon, exhausted_dungeon) and dungeon mutations
(decrement_dungeon, increment_dungeon, eliminate_dungeon) into a new
dungeon.py.  Also extracts treasure inventory operations (_draw,
draw_treasure, replace_treasure) into a new treasure.py.

The former private _blocking_dragon is now public blocking_dragon since
it gains its own dedicated module.

world.py and action.py are now peers — neither imports the other.
Both depend on dungeon.py and treasure.py, which sit at a lower layer.
